+++ client/package.json
@@ -69,7 +69,8 @@
     "postcss": "^8.4.14",
     "prettier": "^2.5.1",
     "vitest": "^0.29.1",
-    "vue-composable-tester": "^0.1.3"
+    "vue-composable-tester": "^0.1.3",
+    "vue-router-mock": "^0.2.0"
   },
   "engines": {
     "node": "^18 || ^16 || ^14.19",
