+++ client/src/components/PublicationStyleCriteria.vitest.spec.js
@@ -2,7 +2,7 @@ import { installQuasarPlugin } from "@quasar/quasar-app-extension-testing-unit-v
 import { ApolloClients } from "@vue/apollo-composable"
 import { mount } from "@vue/test-utils"
 import flushPromises from "flush-promises"
-import { createMockClient } from "mock-apollo-client"
+import { createMockClient } from "test/vitest/apolloClient"
 import {
   CREATE_PUBLICATION_STYLE_CRITERIA,
   DELETE_PUBLICATION_STYLE_CRITERIA,
