+++ client/src/components/atoms/EmailVerificationSendButton.vitest.spec.js
@@ -1,7 +1,7 @@
 import { installQuasarPlugin } from "@quasar/quasar-app-extension-testing-unit-vitest"
 import { mount } from "@vue/test-utils"
 import flushPromises from "flush-promises"
-import { createMockClient } from "mock-apollo-client"
+import { createMockClient } from "test/vitest/apolloClient"
 import { SEND_VERIFY_EMAIL } from "src/graphql/mutations"
 import EmailVerificationSendButton from "./EmailVerificationSendButton.vue"
