+++ client/src/pages/Admin/UsersIndex.vitest.spec.js
@@ -2,7 +2,7 @@ import { installQuasarPlugin } from "@quasar/quasar-app-extension-testing-unit-v
 import { ApolloClients } from "@vue/apollo-composable"
 import { mount } from "@vue/test-utils"
 import flushPromises from "flush-promises"
-import { createMockClient } from "mock-apollo-client"
+import { createMockClient } from "test/vitest/apolloClient"
 import { GET_USERS } from "../../graphql/queries"
 import UsersIndexPage from "./UsersIndex.vue"
