+++ client/test/vitest/apolloClient.js
@@ -1,0 +1,4 @@
+import { createMockClient as createClient } from "mock-apollo-client"
+
+export const createMockClient = (opts) => createClient({...opts, connectToDevTools: false})
+
