+++ client/test/vitest/mockedPlugins/index.js
@@ -1,0 +1,1 @@
+export { default as Dialog } from './Dialog'
