+++ client/test/vitest/setup-file.js
@@ -2,6 +2,7 @@
 import { config } from '@vue/test-utils'
 import { vi } from 'vitest'
 config.global.mocks = {
+  ...config.global.mocks,
   $t: (token) => token,
 }
